Use the new StaticLayout.Builder API on 23+
Summary:
This uses the new StaticLayout.Builder API on 23+.
This allows us to expose break strategy, hyphenation and indentation for the layouts.
Also, bumped the proxy to 1.1 as new API is exposed now.

Test Plan:
Tested the getters and setters.
Tested the new API in a sample app and it works as expected.