Catch NPE from Samsung devices < M when checking if layout is boring
Summary:
On Samsung devices < M there is a possible NPE due to the lack of a
null check within TextLine in Samsung modified versions of Android.

There's not much we can do since we don't necessarily control the
FontMetricsInt object that is causing the NPE, so we just ignore
the exception since its not lethal. In the worst case we will
create a StaticLayout instead of a BoringLayout

Differential Revision: D6991741

fbshipit-source-id: 59e00966e7982f683195e9e5edc4342bff2409e0