Avoid getting the line height of the -1st line
Summary:
If getLineCount() returns 0 (which it can and does from time to time) then we request the -1st line which leads to an ArrayIndexOutOfBoundsException.

This is because (on Android 4.4.2) StaticLayout multiplies the requested line with the mColumns field, which is either 3 or 5, with our requested line number and adds the TOP constant, which is 1. This result is then used as an index into an array, and the only way for it to go negative, as we're observing, is for us to pass in -1, or when getLineCount() returns 0.

Reviewed By: sjkirby

Differential Revision: D10186572

fbshipit-source-id: bccd4339e5064815027c431cecda9e2eaa9fe5e2