Fix layoutpreview4idea build with buck2
Summary:
Somehow missed this earlier...

buck2 doesn't implement `android_aar` rule yet. So I built the aar with buck1 and check it in for now.

This is definitely a hack and authors shouldn't be expected to update this aar manually - open to suggestions on a better fix in the meantime.

Reviewed By: hick209

Differential Revision: D44890940

fbshipit-source-id: 5b55d68723bbed96c58f7433e8cf7726d99815ab