Update StaticLayoutHelper params that received null value to Nullable
Summary:
Runtime plugin showed null ellipsize, leftIndents, and rightIndents passed into make.

## Context
We are in the process of making the entire codebase Nullsafe. This diff relies on a runtime codemod to track when null is passed into a parameter that is not marked as Nullable in java. Based on runtime data, we can be 100% certain that this parameter needs to be Nullable

Reviewed By: jocelynluizzi13

Differential Revision: D60586639

fbshipit-source-id: e2321e611ab96d1a9c4e77b9087fa6865b3794b2